src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "FairFlowAI"
authors = ["Juan José Serrano Gutiérrez <juanjo.jjserra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
aif360 = "0.6.0"
fairlearn = "^0.9.0"
Jinja2 = "^3.0.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
python = ">=3.8,<3.12"
pyyaml = "^5.0.0"
scikit-learn = "^1.2.0"
statsmodels = "^0.14.0"
tempeh = "^0.1.0"
tensorflow = "2.10.0"
tensorflow-io-gcs-filesystem = "^0.30.0"

[tool.poetry.dev-dependencies]
ipykernel = "^6.0.0"
tabulate = "^0.9.0"
matplotlib = "^3.5.0"
seaborn = "^0.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
